requirements.txt
# Core
fastapi==0.135.1
uvicorn==0.44.0
python-dotenv==1.2.2

# Google APIs
google-api-python-client==2.193.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0

# Gemini multimodal
google-genai==1.67.0

# Data validation
pydantic==2.12.5

# HTTP client
httpx==0.28.1

# Rate limiting
slowapi==0.1.9

# Scheduler
apscheduler==3.11.2

# Anthropic (Claude Opus 4.6 — assistant reasoning)
anthropic==0.84.0

# Pinecone (vector DB for course knowledge RAG)
pinecone==8.1.0

# Mem0 (personal memory for WhatsApp assistant)
mem0ai==2.0.0
langchain-neo4j>=0.8.0            # Required by mem0 graph memory (Neo4j AuraDB)
rank-bm25>=0.2.2                  # Required by mem0 graph memory (BM25 reranking)

# Transitive deps pinned to clear Dependabot CVE alerts
# (mem0ai/langchain pull these; locking prevents regressions)
cryptography==46.0.7              # CVE-2026-34073, CVE-2026-39892
pypdf==6.10.2                     # 5 CVEs fixed vs 6.9.1
langchain-core==1.3.1             # CVE-2026-40087
langsmith==0.7.35                 # GHSA-rr7j-v2q5-chgv
requests==2.33.1                  # CVE-2026-25645
pygments==2.20.0                  # CVE-2026-4539

# Development (install via: pip install -r requirements-dev.txt)
# pytest, pytest-asyncio, pytest-cov, ruff
